feat: perf improve on comp calc
